requirements.txt
python-telegram-bot==20.0b0
pandas
requests
beautifulsoup4
matplotlib
fpdf
openpyxl
